fix(ui):为 SobIntroPopup 的 LazyColumn 添加嵌套滚动支持

+++ app/src/main/java/cn/cqautotest/sunnybeach/ui/activity/CreationCenterActivity.kt
@@ -133,13 +133,11 @@ class CreationCenterActivity : AppActivity() {
                     super.onCreate()
                     findViewById<ComposeView>(R.id.compose_view)
                         .setContent {
-                            val maxHeight = remember {
-                                (ScreenUtils.getScreenHeight() / 4f * 3f).pxToDp()
-                            }
+                            val maxHeight = remember { (ScreenUtils.getScreenHeight() / 4f * 3f).pxToDp().toFloat() }
                             SobIntroPopup(
                                 modifier = Modifier
                                     .fillMaxWidth()
-                                    .heightIn(max = Dp(maxHeight.toFloat()))
+                                    .heightIn(max = Dp(maxHeight))
                             )
                         }
                 }
